blockchain/requirements.txt
web3
python-dotenv
py-solc-x
